+++ composition/src/federation/utils.ts
@@ -1,6 +1,7 @@
-import { DocumentNode, GraphQLSchema } from 'graphql';
+import { DocumentNode, FieldDefinitionNode, GraphQLSchema, Kind } from 'graphql';
 import { FieldConfiguration } from '../router-configuration/router-configuration';
 import { SubgraphConfig } from '../subgraph/subgraph';
+import { DefinitionWithFieldsData } from '../schema-building/type-definition-data';
 
 export type FederationResultContainer = {
   errors?: Error[];
@@ -11,8 +12,16 @@ export type FederationResultContainer = {
 export type FederationResult = {
   fieldConfigurations: FieldConfiguration[];
   federatedGraphAST: DocumentNode;
+  federatedGraphClientSchema: GraphQLSchema;
   federatedGraphSchema: GraphQLSchema;
   subgraphConfigBySubgraphName: Map<string, SubgraphConfig>;
+};
+
+export type FederationResultContainerWithContracts = {
+  errors?: Error[];
+  federationResult?: FederationResult;
+  federationResultContainerByContractName?: Map<string, FederationResultContainer>;
+  warnings?: String[];
 };
 
 export type RootTypeFieldData = {
@@ -21,4 +30,37 @@ export type RootTypeFieldData = {
   path: string;
   typeName: string;
   subgraphs: Set<string>;
+};
+
+export type ParentTagData = {
+  childTagDataByChildName: Map<string, ChildTagData>;
+  tagNames: Set<string>;
+  typeName: string;
+};
+
+export function newParentTagData(typeName: string): ParentTagData {
+  return {
+    childTagDataByChildName: new Map<string, ChildTagData>(),
+    tagNames: new Set<string>(),
+    typeName,
+  };
+}
+
+export type ChildTagData = {
+  name: string;
+  tagNames: Set<string>;
+  tagNamesByArgumentName: Map<string, Set<string>>;
+};
+
+export function newChildTagData(name: string): ChildTagData {
+  return {
+    name,
+    tagNames: new Set<string>(),
+    tagNamesByArgumentName: new Map<string, Set<string>>(),
+  };
+}
+
+export type InterfaceImplementationData = {
+  data: DefinitionWithFieldsData;
+  clientSchemaFieldNodes: FieldDefinitionNode[];
 };