+++ connect/src/wg/cosmo/platform/v1/platform_pb.ts
@@ -2477,6 +2477,16 @@ export class Subgraph extends Message<Subgraph> {
    */
   namespace = "";
 
+  /**
+   * @generated from field: string subscriptionProtocol = 11;
+   */
+  subscriptionProtocol = "";
+
+  /**
+   * @generated from field: optional bool isV2Graph = 12;
+   */
+  isV2Graph?: boolean;
+
   constructor(data?: PartialMessage<Subgraph>) {
     super();
     proto3.util.initPartial(data, this);
@@ -2495,6 +2505,8 @@ export class Subgraph extends Message<Subgraph> {
     { no: 8, name: "subscriptionUrl", kind: "scalar", T: 9 /* ScalarType.STRING */ },
     { no: 9, name: "target_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
     { no: 10, name: "namespace", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 11, name: "subscriptionProtocol", kind: "scalar", T: 9 /* ScalarType.STRING */ },
+    { no: 12, name: "isV2Graph", kind: "scalar", T: 8 /* ScalarType.BOOL */, opt: true },
   ]);
 
   static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Subgraph {