+++ controlplane/src/types/index.ts
@@ -102,6 +102,7 @@ export interface SubgraphDTO {
   readme?: string;
   namespace: string;
   namespaceId: string;
+  isV2Graph?: boolean;
 }
 
 export interface MigrationSubgraph {