+++ shared/test/router.config.test.ts
@@ -2,9 +2,17 @@ import * as fs from 'node:fs';
 import * as path from 'node:path';
 import * as url from 'node:url';
 import { describe, expect, test } from 'vitest';
+import { printSchema } from 'graphql';
+import { federateSubgraphs } from '@wundergraph/composition';
 import { buildRouterConfig, ComposedSubgraph } from '../src';
 import { normalizationFailureError } from '../src/router-config/errors';
-import { federateTestSubgraphs } from './testdata/utils';
+import {
+  federateTestSubgraphs,
+  simpleAccounts,
+  simpleProducts,
+  simpleProductsWithInaccessible,
+  simpleProductsWithTags,
+} from './testdata/utils';
 
 // @ts-ignore-next-line
 const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
@@ -72,6 +80,8 @@ describe('Router Config Builder', () => {
       configurationDataMap: inventorySubgraphConfig!.configurationDataMap,
     };
     const routerConfig = buildRouterConfig({
+      // if the federatedClientSDL is empty, it is not added to the config
+      federatedClientSDL: `type Query {}`,
       fieldConfigurations: [],
       subgraphs: [accounts, products, reviews, inventory],
       // Passed as it is to the router config
@@ -83,6 +93,153 @@ describe('Router Config Builder', () => {
     });
     const out = JSON.stringify(JSON.parse(json), null, 2);
     expect(out).matchSnapshot('router.config.json');
+  });
+
+  test('that the federatedClientSDL property is not propagated if it is empty', () => {
+    const { errors, federationResult } = federateSubgraphs([simpleAccounts, simpleProducts]);
+    expect(errors).toBeUndefined();
+    expect(federationResult).toBeDefined();
+    const accountsSubgraphConfig = federationResult?.subgraphConfigBySubgraphName.get('accounts');
+    expect(accountsSubgraphConfig).toBeDefined();
+    const productsSubgraphConfig = federationResult?.subgraphConfigBySubgraphName.get('products');
+    expect(productsSubgraphConfig).toBeDefined();
+
+    const accounts: ComposedSubgraph = {
+      id: '0',
+      name: 'accounts',
+      sdl: fs.readFileSync(path.join(__dirname, 'testdata', 'simple-accounts.graphql'), {
+        encoding: 'utf8',
+      }),
+      url: 'https://wg-federation-demo-accounts.fly.dev/graphql',
+      subscriptionUrl: '',
+      subscriptionProtocol: 'ws',
+      schema: accountsSubgraphConfig!.schema,
+      configurationDataMap: accountsSubgraphConfig!.configurationDataMap,
+    };
+    const products: ComposedSubgraph = {
+      id: '1',
+      name: 'products',
+      sdl: fs.readFileSync(path.join(__dirname, 'testdata', 'simple-products.graphql'), {
+        encoding: 'utf8',
+      }),
+      url: 'https://wg-federation-demo-products.fly.dev/graphql',
+      subscriptionUrl: '',
+      subscriptionProtocol: 'ws',
+      schema: productsSubgraphConfig!.schema,
+      configurationDataMap: productsSubgraphConfig!.configurationDataMap,
+    };
+    const routerConfig = buildRouterConfig({
+      // if the federatedClientSDL is empty, it is not added to the config
+      federatedClientSDL: printSchema(federationResult!.federatedGraphClientSchema),
+      fieldConfigurations: [],
+      subgraphs: [accounts, products],
+      // Passed as it is to the router config
+      federatedSDL: printSchema(federationResult!.federatedGraphSchema),
+      schemaVersionId: '',
+    });
+    const json = routerConfig.toJsonString({
+      emitDefaultValues: false,
+    });
+    const out = JSON.stringify(JSON.parse(json), null, 2);
+    expect(out).matchSnapshot('router-no-client.config.json');
+  });
+
+  test('that the federatedClientSDL property is propagated if a schema uses the @tag directive', () => {
+    const { errors, federationResult } = federateSubgraphs([simpleAccounts, simpleProductsWithTags]);
+    expect(errors).toBeUndefined();
+    expect(federationResult).toBeDefined();
+    const accountsSubgraphConfig = federationResult?.subgraphConfigBySubgraphName.get('accounts');
+    expect(accountsSubgraphConfig).toBeDefined();
+    const productsSubgraphConfig = federationResult?.subgraphConfigBySubgraphName.get('products');
+    expect(productsSubgraphConfig).toBeDefined();
+
+    const accounts: ComposedSubgraph = {
+      id: '0',
+      name: 'accounts',
+      sdl: fs.readFileSync(path.join(__dirname, 'testdata', 'simple-accounts.graphql'), {
+        encoding: 'utf8',
+      }),
+      url: 'https://wg-federation-demo-accounts.fly.dev/graphql',
+      subscriptionUrl: '',
+      subscriptionProtocol: 'ws',
+      schema: accountsSubgraphConfig!.schema,
+      configurationDataMap: accountsSubgraphConfig!.configurationDataMap,
+    };
+    const products: ComposedSubgraph = {
+      id: '1',
+      name: 'products',
+      sdl: fs.readFileSync(path.join(__dirname, 'testdata', 'simple-products-with-tags.graphql'), {
+        encoding: 'utf8',
+      }),
+      url: 'https://wg-federation-demo-products.fly.dev/graphql',
+      subscriptionUrl: '',
+      subscriptionProtocol: 'ws',
+      schema: productsSubgraphConfig!.schema,
+      configurationDataMap: productsSubgraphConfig!.configurationDataMap,
+    };
+    const routerConfig = buildRouterConfig({
+      // if the federatedClientSDL is empty, it is not added to the config
+      federatedClientSDL: printSchema(federationResult!.federatedGraphClientSchema),
+      fieldConfigurations: [],
+      subgraphs: [accounts, products],
+      // Passed as it is to the router config
+      federatedSDL: printSchema(federationResult!.federatedGraphSchema),
+      schemaVersionId: '',
+    });
+    const json = routerConfig.toJsonString({
+      emitDefaultValues: false,
+    });
+    const out = JSON.stringify(JSON.parse(json), null, 2);
+    expect(out).matchSnapshot('router-with-tags.config.json');
+  });
+
+  test('that the federatedClientSDL property is propagated if a schema uses the @inaccessible directive', () => {
+    const { errors, federationResult } = federateSubgraphs([simpleAccounts, simpleProductsWithInaccessible]);
+    expect(errors).toBeUndefined();
+    expect(federationResult).toBeDefined();
+    const accountsSubgraphConfig = federationResult?.subgraphConfigBySubgraphName.get('accounts');
+    expect(accountsSubgraphConfig).toBeDefined();
+    const productsSubgraphConfig = federationResult?.subgraphConfigBySubgraphName.get('products');
+    expect(productsSubgraphConfig).toBeDefined();
+
+    const accounts: ComposedSubgraph = {
+      id: '0',
+      name: 'accounts',
+      sdl: fs.readFileSync(path.join(__dirname, 'testdata', 'simple-accounts.graphql'), {
+        encoding: 'utf8',
+      }),
+      url: 'https://wg-federation-demo-accounts.fly.dev/graphql',
+      subscriptionUrl: '',
+      subscriptionProtocol: 'ws',
+      schema: accountsSubgraphConfig!.schema,
+      configurationDataMap: accountsSubgraphConfig!.configurationDataMap,
+    };
+    const products: ComposedSubgraph = {
+      id: '1',
+      name: 'products',
+      sdl: fs.readFileSync(path.join(__dirname, 'testdata', 'simple-products-with-inaccessible.graphql'), {
+        encoding: 'utf8',
+      }),
+      url: 'https://wg-federation-demo-products.fly.dev/graphql',
+      subscriptionUrl: '',
+      subscriptionProtocol: 'ws',
+      schema: productsSubgraphConfig!.schema,
+      configurationDataMap: productsSubgraphConfig!.configurationDataMap,
+    };
+    const routerConfig = buildRouterConfig({
+      // if the federatedClientSDL is empty, it is not added to the config
+      federatedClientSDL: printSchema(federationResult!.federatedGraphClientSchema),
+      fieldConfigurations: [],
+      subgraphs: [accounts, products],
+      // Passed as it is to the router config
+      federatedSDL: printSchema(federationResult!.federatedGraphSchema),
+      schemaVersionId: '',
+    });
+    const json = routerConfig.toJsonString({
+      emitDefaultValues: false,
+    });
+    const out = JSON.stringify(JSON.parse(json), null, 2);
+    expect(out).matchSnapshot('router-with-inaccessible.config.json');
   });
 
   test('that the builder config throws an error if normalization has failed', () => {
@@ -99,6 +256,7 @@ describe('Router Config Builder', () => {
     let error;
     try {
       buildRouterConfig({
+        federatedClientSDL: '',
         fieldConfigurations: [],
         subgraphs: [subgraph],
         federatedSDL: '',